requirements.txt
flask
flask_cors
rembg
Pillow
gunicorn
waitress
fastapi
uvicorn
python-multipart
tornado



